moved the version to it's own package to avoid cyclical deps
